BackEnd/requirements.txt
flask
flask-cors
psycopg2-binary
SQLAlchemy
python-dotenv
